Post is working when tradeId is int, should work when changed to long

+++ StockClient/Models/UserTrade.cs
@@ -6,7 +6,7 @@ namespace StockClient.Models
 {
     public class UserTrade
     {
-        public int TradeId { get; set; }
+        public long TradeId { get; set; }
         public string Username { get; set; }
         public string Symbol { get; set; }
         public float Units { get; set; }
